[wts] Rewrite cgi to py for CSP
Impacted tests(approved): new 0, update 35, delete 0
Unit test platform: WTS
Unit test result summary: pass 17, fail 18, block 0

BUG=https://crosswalk-project.org/jira/browse/XWALK-4504
