+++ webapi/tct-csp-w3c-tests/csp-py/csp_ro_connect-src_self.py
@@ -1,10 +1,15 @@
-#!/bin/sh
-echo "Content-Security-Policy-Report-Only: connect-src *"
-echo "X-Content-Security-Policy-Report-Only: connect-src *"
-echo "X-WebKit-CSP-Report-Only: connect-src *"
-echo
-echo '<!DOCTYPE html>
-
+def main(request, response):
+    import simplejson as json
+    f = file('config.json')
+    source = f.read()
+    s = json.JSONDecoder().decode(source)
+    url1 = "http://" + s['host'] + ":" + str(s['ports']['http'][1])
+    url2 = "http://" + s['host'] + ":" + str(s['ports']['http'][0])
+    _CSP = "connect-src 'self'; script-src 'self' 'unsafe-inline'"
+    response.headers.set("Content-Security-Policy", _CSP)
+    response.headers.set("X-Content-Security-Policy", _CSP)
+    response.headers.set("X-WebKit-CSP", _CSP)
+    return """<!DOCTYPE html>
 <!--
 Copyright (c) 2013 Samsung Electronics Co., Ltd.
 
@@ -26,11 +31,11 @@ Authors:
 
 <html>
   <head>
-    <title>CSP Test: csp_ro_connect-src_asterisk</title>
+    <title>CSP Test: csp_ro_connect-src_self_xmlhttprequest</title>
     <link rel="author" title="Samsung" href="http://www.Samsung.com"/>
     <link rel="help" href="http://www.w3.org/TR/2012/CR-CSP-20121115/#connect-src"/>
     <meta name="flags" content=""/>
-    <meta name="assert" content="connect-src *"/>
+    <meta name="assert" content="connect-src 'self'; script-src 'self' 'unsafe-inline'"/>
     <meta charset="utf-8"/>
     <script src="../resources/testharness.js"></script>
     <script src="../resources/testharnessreport.js"></script>
@@ -38,29 +43,23 @@ Authors:
   <body>
     <div id="log"></div>
     <script>
+        var xhr = new XMLHttpRequest();
+
         test(function() {
-            var xhr = new XMLHttpRequest();
             try {
                 xhr.open("GET", "support/csp.js");
             } catch(e) {
-                assert_unreached("should not reach here, got exception: "+e.message);
+                assert_unreached("Should not reach here, exception: " + e.message);
             }
         }, document.title + "_allowed");
 
         test(function() {
-            var xhr = new XMLHttpRequest();
             try {
-                xhr.open("GET", "http://www.w3.org");
+                xhr.open("GET", "http://https://www.tizen.org");
             } catch(e) {
-                assert_unreached("should not reach here, got exception: "+e.message);
+                assert_unreached("Should not reach here, exception: " + e.message);
             }
-        }, document.title + "_allowed_ext");
-
-        function log123(message) {
-            var client = new XMLHttpRequest();
-            client.open("POST", "http://127.0.0.1:8000/log123");
-            client.send(message);
-        }
+        }, document.title + "_blocked");
     </script>
   </body>
-</html> '
+</html>"""
