+++ webapi/tct-csp-w3c-tests/csp-py/csp_ro_media-src_none_audio_blocked_ext.py
@@ -1,10 +1,15 @@
-#!/bin/sh
-echo "Content-Security-Policy-Report-Only: media-src 'none'; script-src 'self' 'unsafe-inline'"
-echo "X-Content-Security-Policy-Report-Only: media-src 'none'; script-src 'self' 'unsafe-inline'"
-echo "X-WebKit-CSP-Report-Only: media-src 'none'; script-src 'self' 'unsafe-inline'"
-echo
-echo '<!DOCTYPE html>
-
+def main(request, response):
+    import simplejson as json
+    f = file('config.json')
+    source = f.read()
+    s = json.JSONDecoder().decode(source)
+    url1 = "http://" + s['host'] + ":" + str(s['ports']['http'][1])
+    url2 = "http://" + s['host'] + ":" + str(s['ports']['http'][0])
+    _CSP = "media-src 'none'; script-src 'self' 'unsafe-inline'"
+    response.headers.set("Content-Security-Policy", _CSP)
+    response.headers.set("X-Content-Security-Policy", _CSP)
+    response.headers.set("X-WebKit-CSP", _CSP)
+    return """<!DOCTYPE html>
 <!--
 Copyright (c) 2013 Samsung Electronics Co., Ltd.
 
@@ -26,7 +31,7 @@ Authors:
 
 <html>
   <head>
-    <title>CSP Test: csp_ro_media-src_none_video_blocked_ext</title>
+    <title>CSP Test: csp_ro_media-src_none_audio_blocked_ext</title>
     <link rel="author" title="Samsung" href="http://www.Samsung.com"/>
     <link rel="help" href="http://www.w3.org/TR/2012/CR-CSP-20121115/#media-src"/>
     <meta name="flags" content=""/>
@@ -37,18 +42,18 @@ Authors:
   </head>
   <body>
     <div id="log"></div>
-    <video id="m"></video>
+    <audio id="m"></audio>
     <script>
         var t = async_test(document.title);
         var m = document.getElementById("m");
-        m.src = "http://127.0.0.1:8083/opt/tct-csp-w3c-tests/csp/support/red-green.theora.ogv";
+        m.src = '""" + s['host'] + """:8083/tests/csp/support/red-green.theora.ogv';
         window.setTimeout(function() {
             t.step(function() {
                 assert_false(m.currentSrc == "",
-                    "video.currentSrc should not be empty even use the not allowed external audio resource in report only mode");
+                    "audio.currentSrc should not be empty if use the external audio resource when media-src is none in report only mode.");
             });
             t.done();
         }, 0);
     </script>
   </body>
-</html> '
+</html>"""
