+++ webapi/tct-csp-w3c-tests/csp-py/csp_ro_style-src_cross-origin.py
@@ -1,10 +1,15 @@
-#!/bin/sh
-echo "Content-Security-Policy-Report-Only:default *;style-src 'unsafe-inline'"
-echo "X-Content-Security-Policy-Report-Only:default *;style-src 'unsafe-inline'"
-echo "X-WebKit-CSP-Report-Only:default *;style-src 'unsafe-inline'"
-echo
-echo '<!DOCTYPE html>
-
+def main(request, response):
+    import simplejson as json
+    f = file('config.json')
+    source = f.read()
+    s = json.JSONDecoder().decode(source)
+    url1 = "http://" + s['host'] + ":" + str(s['ports']['http'][1])
+    url2 = "http://" + s['host'] + ":" + str(s['ports']['http'][0])
+    _CSP = "style-src " + url1
+    response.headers.set("Content-Security-Policy", _CSP)
+    response.headers.set("X-Content-Security-Policy", _CSP)
+    response.headers.set("X-WebKit-CSP", _CSP)
+    return """<!DOCTYPE html>
 <!--
 Copyright (c) 2013 Samsung Electronics Co., Ltd.
 
@@ -26,15 +31,16 @@ Authors:
 
 <html>
   <head>
-    <title>CSP Test: csp_ro_style-src_unsafe-inline</title>
+    <title>CSP Test: csp_ro_style-src_cross-origin</title>
     <link rel="author" title="Samsung" href="http://www.Samsung.com"/>
     <link rel="help" href="http://www.w3.org/TR/2012/CR-CSP-20121115/#style-src"/>
     <meta name="flags" content=""/>
-    <meta name="assert" content="default *;style-src 'unsafe-inline'"/>
+    <meta name="assert" content='style-src """ + url1 + """'/>
     <meta charset="utf-8"/>
     <script src="../resources/testharness.js"></script>
     <script src="../resources/testharnessreport.js"></script>
-    <link rel="stylesheet" type="text/css" href="http://127.0.0.1:8081/opt/tct-csp-w3c-tests/csp/support/canvas-index.css"/>
+    <link rel="stylesheet" type="text/css" href='""" + url1 + """/tests/csp/support/canvas-index.css'/>
+    <link rel="stylesheet" type="text/css" href='""" + url2 + """/tests/csp/support/a-green.css'/>
     <link rel="stylesheet" type="text/css" href="support/blue-100x100.css"/>
     <style>
       #test-green {
@@ -45,20 +51,27 @@ Authors:
   <body>
     <div id="log"></div>
     <div id="test-blue"></div>
+    <h3>ext-css:""" + url1 + """/tests/csp/support/canvas-index.css</h3>
+    <div id="test-ext-a" class="a"></div>
     <div id="test-green"></div>
-    <h3>ext-css:http://127.0.0.1:8081/opt/tct-csp-w3c-tests/csp/support/canvas-index.css</h3>
     <script>
         test(function() {
-            var div = document.querySelector("h3");
-            var fix = getComputedStyle(div)["display"];
-            assert_equals(fix, "inline", "style setted correctly");
-        }, document.title + "_blocked_ext");
+            var div = document.querySelector("#test-ext-a");
+            var fix = getComputedStyle(div)["color"];
+            assert_equals(fix, "rgb(0, 128, 0)", "style setted correctly");
+        }, document.title + "_blocked");
 
         test(function() {
             var div = document.querySelector("#test-blue");
             var fix = getComputedStyle(div)["backgroundColor"];
             assert_equals(fix, "rgb(0, 0, 255)", "style setted correctly");
-        }, document.title + "_bloced_int");
+        }, document.title + "_blocked_int");
+
+        test(function() {
+            var div = document.querySelector("#test-green");
+            var fix = getComputedStyle(div)["backgroundColor"];
+            assert_equals(fix, "rgb(0, 128, 0)", "style setted correctly");
+        }, document.title + "_blocked_inline");
     </script>
   </body>
-</html> '
+</html>"""
