+++ webapi/tct-csp-w3c-tests/csp-py/csp_ro_style-src_self.py
@@ -1,10 +1,15 @@
-#!/bin/sh
-echo "Content-Security-Policy-Report-Only:style-src *"
-echo "X-Content-Security-Policy-Report-Only:style-src *"
-echo "X-WebKit-CSP-Report-Only:style-src *"
-echo
-echo '<!DOCTYPE html>
-
+def main(request, response):
+    import simplejson as json
+    f = file('config.json')
+    source = f.read()
+    s = json.JSONDecoder().decode(source)
+    url1 = "http://" + s['host'] + ":" + str(s['ports']['http'][1])
+    url2 = "http://" + s['host'] + ":" + str(s['ports']['http'][0])
+    _CSP = "style-src 'self'"
+    response.headers.set("Content-Security-Policy", _CSP)
+    response.headers.set("X-Content-Security-Policy", _CSP)
+    response.headers.set("X-WebKit-CSP", _CSP)
+    return """<!DOCTYPE html>
 <!--
 Copyright (c) 2013 Samsung Electronics Co., Ltd.
 
@@ -26,15 +31,15 @@ Authors:
 
 <html>
   <head>
-    <title>CSP Test: csp_ro_style-src_asterisk</title>
-    <link rel="author" title="Samsung" href="http://www.Samsung.com/"/>
+    <title>CSP Test: csp_ro_style-src_self</title>
+    <link rel="author" title="Samsung" href="http://www.Samsung.com"/>
     <link rel="help" href="http://www.w3.org/TR/2012/CR-CSP-20121115/#style-src"/>
     <meta name="flags" content=""/>
-    <meta name="assert" content="style-src *"/>
+    <meta name="assert" content="style-src 'self'"/>
     <meta charset="utf-8"/>
     <script src="../resources/testharness.js"></script>
     <script src="../resources/testharnessreport.js"></script>
-    <link rel="stylesheet" type="text/css" href="http://127.0.0.1:8081/opt/tct-csp-w3c-tests/csp/support/canvas-index.css"/>
+    <link rel="stylesheet" type="text/css" href='""" + url1 + """/tests/csp/support/canvas-index.css'/>
     <link rel="stylesheet" type="text/css" href="support/blue-100x100.css"/>
     <style>
       #test-green {
@@ -46,8 +51,14 @@ Authors:
     <div id="log"></div>
     <div id="test-blue"></div>
     <div id="test-green"></div>
-    <h3>ext-css:http://127.0.0.1:8081/opt/tct-csp-w3c-tests/csp/support/canvas-index.css</h3>
+    <h3>ext-css:""" + url1 + """/tests/csp/support/canvas-index.css</h3>
     <script>
+        test(function() {
+            var div = document.querySelector("h3");
+            var fix = getComputedStyle(div)["display"];
+            assert_equals(fix, "inline", "style setted correctly");
+        }, document.title + "_blocked");
+
         test(function() {
             var div = document.querySelector("#test-green");
             var fix = getComputedStyle(div)["backgroundColor"];
@@ -55,4 +66,4 @@ Authors:
         }, document.title + "_blocked_inline");
     </script>
   </body>
-</html> '
+</html>"""
